Add files via upload
This is the creation of the initial table that will house the historical data. The date column is a text data type, because the date data is in 'dd-mm-yyyy' format. In the next file, the date column data type will be converted to a date type, and a new table will be created using the exported updated data.